pyproject.toml
[tool.poetry]
name = "pdfchat-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
tqdm = "^4.66.1"
unstructured = {extras = ["pdf"], version = "^0.12.2"}
python-dotenv = "^1.0.0"
openai = "^1.9.0"
psycopg = {extras = ["binary"], version = "^3.2.4"}
pgvector = "^0.2.4"
httpx = "0.27.2"
langchain-community = "^0.3.16"
langchain-openai = "^0.3.3"
langchain-experimental = "^0.3.4"
langchain-postgres = "^0.0.13"
asyncpg = "^0.30.0"
llama-index-embeddings-openai = "^0.3.1"
llama-index-readers-file = "^0.4.4"
langchain = "^0.3.18"
langchain-ollama = "^0.2.3"
sentence-transformers = "^3.4.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"